move content blocks from includes/templates/ to new includes/content/ directory
