+++ catalog/checkout_shipping_address.php
@@ -297,7 +297,7 @@
 
   <div class="alert alert-info"><?php echo TEXT_CREATE_NEW_SHIPPING_ADDRESS; ?></div>
 
-  <?php require('includes/modules/checkout_new_address.php'); ?>
+  <?php require('includes/content/checkout_new_address.php'); ?>
 
 <?php
   }
