+++ catalog/product_info.php
@@ -243,7 +243,7 @@
   </div>
 
 <?php
-    include('includes/modules/also_purchased_products.php');
+    include('includes/content/also_purchased_products.php');
 
     if ( $Qproduct->valueInt('manufacturers_id') > 0 ) {
       $Qm = $OSCOM_Db->get('manufacturers', 'manufacturers_name', ['manufacturers_id' => $Qproduct->valueInt('manufacturers_id')]);
