+++ catalog/specials.php
@@ -108,7 +108,7 @@
     }
   }
 
-  include('includes/modules/product_listing.php');
+  include('includes/content/product_listing.php');
 
   require($oscTemplate->getFile('template_bottom.php'));
   require('includes/application_bottom.php');
